⭐Change flight detail screen Ui
