+++ lib/core/components/enum/tic_type_enum.dart
@@ -6,29 +6,17 @@ enum TicTypeEnum {
   businessClass,
   firstClass;
 
-  String get displayValue {
-    switch (this) {
-      case TicTypeEnum.economyClass:
-        return "Economy Class";
-      case TicTypeEnum.premiumEconomyClass:
-        return "Premium Economy Class";
-      case TicTypeEnum.businessClass:
-        return "Business Class";
-      case TicTypeEnum.firstClass:
-        return "First Class";
-    }
-  }
+  String get displayValue => switch (this) {
+        TicTypeEnum.economyClass => "Economy Class",
+        TicTypeEnum.premiumEconomyClass => "Premium Class",
+        TicTypeEnum.businessClass => "Business Class",
+        _ => "First Class"
+      };
 
-  Color get colorType {
-    switch (this) {
-      case TicTypeEnum.economyClass:
-        return Colors.blue;
-      case TicTypeEnum.premiumEconomyClass:
-        return Colors.green;
-      case TicTypeEnum.businessClass:
-        return Colors.red;
-      case TicTypeEnum.firstClass:
-        return Colors.yellow;
-    }
-  }
+  Color get colorType => switch (this) {
+        TicTypeEnum.economyClass => Colors.blue,
+        TicTypeEnum.premiumEconomyClass => Colors.green,
+        TicTypeEnum.businessClass => Colors.red,
+        _ => Colors.yellow,
+      };
 }
