+++ lib/core/components/enum/type_employee_enum.dart
@@ -4,16 +4,10 @@ enum TypeEmployeeEnum {
   pilot,
   firstOfficer;
 
-  String get displayValue {
-    switch (this) {
-      case TypeEmployeeEnum.flightAttendant:
-        return "Flight attendant";
-      case TypeEmployeeEnum.cabinCrew:
-        return "Cabin crew";
-      case TypeEmployeeEnum.pilot:
-        return "Pilot";
-      case TypeEmployeeEnum.firstOfficer:
-        return "First Officer";
-    }
-  }
+  String get displayValue => switch (this) {
+        TypeEmployeeEnum.flightAttendant => "Flight attendant",
+        TypeEmployeeEnum.cabinCrew => "Cabin crew",
+        TypeEmployeeEnum.pilot => "Pilot",
+        _ => "First Officer",
+      };
 }
