+++ lib/l10n/intl_en.arb
@@ -244,5 +244,7 @@
     "addReview": "Add review",
     "rateYourStay": "Rate your stay",
     "yourReview": "Your Review",
-    "gallery": "Gallery"
+    "gallery": "Gallery",
+    "selectScott": "Select Scott",
+    "airlines": "Airlines"
 }
